pyproject.toml
[project]
name = "layrd-sync-agent"
version = "0.5.0"
description = "Layrd Document Sync Agent — watches local/network folders and uploads documents to Layrd"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "pystray>=0.19",
    "Pillow>=10.0",
    "apscheduler>=3.10",
    "platformdirs>=4.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pyinstaller>=6.0",
]

[project.scripts]
layrd-sync = "layrd_sync.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["layrd_sync"]
